portal-aulas-api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
django==4.0.10
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==23.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
psycopg==3.1.8
psycopg2-binary==2.9.5
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
